requirements.txt
# Web framework
Flask==3.1.0
Werkzeug==3.1.3

# Deep learning
torch==2.6.0
torchaudio==2.6.0

# Audio processing
librosa==0.10.2
soundfile==0.13.1
demucs==4.0.1
pyannote.audio==3.1.1
pydub==0.25.1

# Scientific computing
numpy==1.24.3
scipy==1.15.1

# File handling and utilities
click==8.1.8

# Machine learning
scikit-learn==1.6.1

# Additional dependencies
python-dotenv==1.0.0
